==> scripts/01_data_cleaning.Rmd <==
---
title: "SBC LTER Data"
author: "Mary Fisher"
date: "February 24, 2020"
output: 
  html_document:
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(here)
library(vegan)
library(lubridate)
```

## Biological Communities Data

### Raw Data
Abundance and size of Giant Kelp (Macrocystis Pyrifera), ongoing since 2000. 
```{r}
kelp <- read.csv(here::here("data","raw","Annual_Kelp_All_Years_20181127.csv")) %>%
  clean_names() %>%
  remove_empty("rows") %>%
  na_if(-99999)
colnames(kelp); dim(kelp)
```
<br>

Abundance and size of fish at kelp forest monitoring sites, ongoing since 2000. 
```{r}
fish <- read.csv(here::here("data","raw","Annual_fish_comb_20181127.csv")) %>%
  clean_names() %>%
  remove_empty("rows") %>%
  na_if(-99999)
colnames(fish); dim(fish)
```
<br>


Abundance of algae and invertebrates at kelp forest monitoring sites, ongoing since 2000. 
```{r}
invert <- read.csv(here::here("data","raw","Annual_Quad_Swath_All_Years_20181127.csv")) %>%
  clean_names() %>%
  remove_empty("rows") %>%
  na_if(-99999)
colnames(invert); dim(invert)
```
<br>

### Processed Data Set

Is each site represented in each data set?
```{r}
all(unique(kelp$site) %in% unique(fish$site))
all(unique(kelp$site) %in% unique(invert$site))
```
<br>

How many unique species / taxonomic orders are in each data set?
```{r}
unique(kelp$sp_code)
unique(fish$sp_code); unique(fish$taxon_order)
unique(invert$sp_code); unique(invert$taxon_order)
```
<br>

#### Kelp Data

*LTER site Abstract:* These data describe the abundance of giant kelp, Macrocystis pyrifera, and are part of the SBCLTER kelp forest monitoring program. The study was initiated in 2000, in the Santa Barbara Channel, California, USA, and this dataset is updated once per year. The number and size (number of fronds and diameter of holdfasts) of plants were recorded along permanent transects at nine reef sites located along the mainland coast of the Santa Barbara Channel and at two sites on the north side Santa Cruz Island. These sites reflect several oceanographic regimes in the channel and vary in distance from sources of terrestrial runoff. The time period of data collection varied among the 11 kelp forest sites. Sampling at BULL, CARP, and NAPL began in 2000, sampling at the other 6 mainland sites (AHND, AQUE, IVEE, GOLB, ABUR, MOHK) began in 2001 (transects 3, 5, 6, 7, 8 at IVEE were added in 2011). 

I'm going to grab the data for the mainland sites, for 2001 onward. 
```{r}
kelp_thin <- dplyr::select(kelp, year, month, date,site,transect,obs_code, quad,side,survey,sp_code,scientific_name,group,fronds,hld_diam,area) %>%
  filter(year >= 2001) %>%
  filter(!(site == "IVEE" & transect %in% c(3,5,6,7,8)))
head(kelp_thin)
```
<br>

Now I have to summarise the size / abundance of kelp at each site. The data was collected in quadrats along permanent transects, so each transect is a replicate. A transect has 2 sections ("quad"), 0-20m and 20-40m, and 2 sides ("side"), inshore and offshore. For abundance, I want to first calculate the mean and maximum fronds / m of the transect. For size, I want to calculate the mean  and maximum holdfast diameter per transect.

There are four species codes in the kelp data set, representing different life stages: MAPY (adult giant kelp), MH (giant kelp holdfast), MJ (juvenile giant kelp), and MPS (stipe giant kelp). Two of these (MAPY, MPS) are present in the data set. I'll just calculate the total abundance and mean size for adults.
```{r}
kelp_abund <- kelp_thin %>%
  filter(sp_code=="MAPY") %>%
  #calculate fronds / m^2
  group_by(year, month, site,transect,scientific_name,group,quad,side) %>%
  summarise(abund=sum(fronds,na.rm=TRUE)/sum(area)) %>%
  #mean of fronds / m
  group_by(year, month, site,transect,scientific_name,group) %>%
  summarise(MAPY_mean_abund=mean(abund,na.rm=TRUE),MAPY_max_abund=max(abund,na.rm=TRUE))

kelp_size <- kelp_thin %>%
  filter(sp_code=="MAPY") %>%
  #mean of holdfast diameter, per transect
  group_by(year, month, site,transect,scientific_name,group) %>%
  summarise(MAPY_mean_size=mean(hld_diam,na.rm=TRUE),MAPY_max_size=max(hld_diam,na.rm=TRUE)) %>%
  na_if(NaN) %>%
  na_if(-Inf)

kelp_means <- full_join(kelp_abund,kelp_size,by=c("year", "month", "site","transect","scientific_name","group"))

head(kelp_means)
```
<br>

Write out the new data
```{r}
write.csv(kelp_means,here::here("data","Annual_Kelp_MeanSize_MeanAbund_2001-2018.csv"),row.names=FALSE)
```
<br>


#### Fish Data
*LTER site Abstract:* These data describe the abundance and size of fish species as part of SBCLTER's kelp forest monitoring program to track long-term patterns in species abundance and diversity. The abundance and size of all taxa of resident kelp forest fish encountered along permanent transects are recorded at nine reef sites located along the mainland coast of the Santa Barbara Channel and at two sites on the north side of Santa Cruz Island. These sites reflect several oceanographic regimes in the channel and vary in distance from sources of terrestrial runoff. In these surveys, fish were counted in a 40x2m benthic quadrat ("cryptic fish"), or in the water parcel 0-2m off the bottom over the same area ("fish"). The annual benthic fish community survey which was conducted on 11 reefs once a year around late July or early August. The time period of data collection for the annual benthic fish community survey varied among the 11 kelp forest sites. Sampling at BULL, CARP, and NAPL began in 2000, sampling at the other 6 mainland sites (AHND, AQUE, IVEE, GOLB, ABUR, MOHK) began in 2001 (transects 3, 5, 6, 7, 8 at IVEE were added in 2011). 


I'm going to grab the data for the mainland sites, for 2001 onward. 

```{r}
fish_thin <- dplyr::select(fish, year, month, date,site,transect, quad,side,survey,taxon_order, sp_code,scientific_name,group,size,count,area) %>%
  filter(year >= 2001) %>%
  filter(!(site == "IVEE" & transect %in% c(3,5,6,7,8)))
head(fish_thin)
```
<br>


I'll summarize this data set using (1) species richness, (2) species abundance, and (3) size structure, per side / quad / transect.

*Species richness* is the number of species at a site. Only the "cryptic fish" survey method - observing fish sitting on the bottom using a quadrat - includes a side and quadrat measurement for the transect. Under the assumption that these fish did not move during the survey (since they are usually small with limited ranges), I'm going to sum the number of species across quadrats / sides. Just in case, I'll also calculate species richness with Menhinick's index - the number of species (n) divided by the square-root of the total number of individuals (N).
*Species abundance* is just the total number of individual fish at a site. Since survey methods were consistent over time, we can use a raw abundance measure instead of a density estimate.
```{r}
sp.richness <- fish_thin %>%
  group_by(year,month,date,site,transect,survey) %>%
  filter(count != 0) %>%
  summarise(sp.richness = length(unique(sp_code)), sp.mrichness = sp.richness / sqrt(sum(count)))
order.richness <- fish_thin %>%
  group_by(year,month,date,site,transect,survey) %>%
  filter(count != 0) %>%
  summarise(order.richness = length(unique(taxon_order)), order.mrichness = order.richness / sqrt(sum(count)))

abundance <- fish_thin %>%
   group_by(year,month,date,site,transect,survey) %>%
  summarise(abundance = sum(count))
```
<br>

*Species diversity* can be quantified in a couple of ways, but the most common is the Shannon Index: the uncertainty associated with predicting the identity of a new taxa given number of taxa and evenness in abundances of individuals within each taxa. Shannon's Index can range from 0 to 5, which may make it less appealing to work with than the Simpson's Index, which is a proportion that ranges from 0 to 1. Simpson's Index is the probability that two individuals drawn at random from an infinitely large community will be different species, and is less sensitive to rare species than the Shannon Index. So I'll calculate both here.  
```{r}
fish_wide <- fish_thin %>%
  # get total count for each species, across quadrats / sides when appropriate
  group_by(year,month,date,site,transect,survey,sp_code) %>%
  summarise(total_count = sum(count)) %>%
  # make data frame longer so each species has its own column
  pivot_wider(id_cols=c(year,month,date,site,transect,survey), names_from=sp_code, values_from=total_count) %>% ungroup()
# grab just the data and just the ids
fish_wide_data <- dplyr::select(fish_wide,-c(year,month,date,site,transect,survey))
fish_wide_data[is.na(fish_wide_data)] <- 0
fish_wide_ids <- dplyr::select(fish_wide,c(year,month,date,site,transect,survey))
# calculate diversity
simpson.div <- diversity(fish_wide_data,index="simpson")
shannon.div <- diversity(fish_wide_data)
  
sp.diversity <- fish_wide_ids %>%
  mutate(div.simpson = simpson.div, div.shannon = shannon.div)
```
<br>

In case species-level data isn't reliable, I'll also calculate order-level diversity.
```{r}
fish_wide <- fish_thin %>%
  # get total count for each species, across quadrats / sides when appropriate
  group_by(year,month,date,site,transect,survey,taxon_order) %>%
  summarise(total_count = sum(count)) %>%
  # make data frame longer so each species has its own column
  pivot_wider(id_cols=c(year,month,date,site,transect,survey), names_from=taxon_order, values_from=total_count) %>% ungroup()
# grab just the data and just the ids
fish_wide_data <- dplyr::select(fish_wide,-c(year,month,date,site,transect,survey))
fish_wide_data[is.na(fish_wide_data)] <- 0
fish_wide_ids <- dplyr::select(fish_wide,c(year,month,date,site,transect,survey))
# calculate diversity
simpson.div <- diversity(fish_wide_data,index="simpson")
shannon.div <- diversity(fish_wide_data)
  
diversity <- fish_wide_ids %>%
  mutate(order.div.simpson = simpson.div, order.div.shannon = shannon.div) %>%
  left_join(sp.diversity,by=c("year","month","date","site","transect","survey"))
```
<br>

Combine the data sets and write out: 
```{r}
fish_out <- left_join(sp.richness,abundance,by=c("year","month","date","site","transect","survey")) %>%
  left_join(order.richness, by=c("year","month","date","site","transect","survey")) %>%
  left_join(diversity, by=c("year","month","date","site","transect","survey"))
head(fish_out)
write.csv(fish_out, here::here("data","Annual_Fish_RichnessAbundDiversity_2001-2018.csv"),row.names=FALSE)
```
<br>

#### Invertebrate Data

*LTER site Abstract:* These data describe the density and size of a select group of about 50 species of benthic invertebrates and understory algae in fixed plots (either 1m x 1m or 20m x 1m) along permanent transects. These data are part of SBCLTER's kelp forest monitoring program to track long-term patterns in species abundance and diversity of reef-associated organisms in the Santa Barbara Channel, California, USA. The sampling locations in this dataset are at nine reef sites along the mainland coast of the Santa Barbara Channel and at two sites on the north side of Santa Cruz Island. These sites reflect several oceanographic regimes in the channel and vary in distance from sources of terrestrial runoff. Sampling began in 2000, and these data are updated annually. The time period of data collection varied among the 11 kelp forest sites. Sampling at BULL, CARP, and NAPL began in 2000, sampling at the other 6 mainland sites (AHND, AQUE, IVEE, GOLB, ABUR, MOHK) began in 2001 (transects 3, 5, 6, 7, 8 at IVEE were added in 2011). Data collection at the two Santa Cruz Island sites (SCTW and SCDI) began in 2004. See Methods for more information.  


I'm going to grab the data for invertebrates from the mainland sites, for 2001 onward. 

```{r}
invert_thin <- dplyr::select(invert, year, month, date,site,transect, quad,side,survey,taxon_order, sp_code,scientific_name,group,size,count,area) %>%
  filter(group=="INVERT") %>%
  filter(year >= 2001) %>%
  filter(!(site == "IVEE" & transect %in% c(3,5,6,7,8)))
head(invert_thin)
```
<br>

I'll then use essentially the same metrics as calculated above, for the fish data set. I'm also going to add a column with urchin abundance (urchins / m^2). 

Species richness and abundance
```{r}
sp.richness <- invert_thin %>%
  group_by(year,month,date,site,transect,survey) %>%
  filter(count != 0) %>%
  summarise(sp.richness = length(unique(sp_code)), sp.mrichness = sp.richness / sqrt(sum(count)))

abundance <- invert_thin %>%
   group_by(year,month,date,site,transect,survey) %>%
  summarise(abundance = sum(count))
```
<br>

Species diversity
```{r}
invert_wide <- invert_thin %>%
  # get total count for each species, across quadrats / sides when appropriate
  group_by(year,month,date,site,transect,survey,sp_code) %>%
  summarise(total_count = sum(count)) %>%
  # make data frame longer so each species has its own column
  pivot_wider(id_cols=c(year,month,date,site,transect,survey), names_from=sp_code, values_from=total_count) %>% ungroup()
# grab just the data and just the ids
invert_wide_data <- dplyr::select(invert_wide,-c(year,month,date,site,transect,survey))
invert_wide_data[is.na(invert_wide_data)] <- 0
invert_wide_ids <- dplyr::select(invert_wide,c(year,month,date,site,transect,survey))
# calculate diversity
simpson.div <- diversity(invert_wide_data,index="simpson")
shannon.div <- diversity(invert_wide_data)
  
sp.diversity <- invert_wide_ids %>%
  mutate(div.simpson = simpson.div, div.shannon = shannon.div)
```
<br>

Taxon (order) richness and diversity
```{r}
order.richness <- invert_thin %>%
  group_by(year,month,date,site,transect,survey) %>%
  filter(count != 0) %>%
  summarise(order.richness = length(unique(taxon_order)), order.mrichness = order.richness / sqrt(sum(count)))

invert_wide <- invert_thin %>%
  # get total count for each species, across quadrats / sides when appropriate
  group_by(year,month,date,site,transect,survey,taxon_order) %>%
  summarise(total_count = sum(count)) %>%
  # make data frame longer so each species has its own column
  pivot_wider(id_cols=c(year,month,date,site,transect,survey), names_from=taxon_order, values_from=total_count) %>% ungroup()
# grab just the data and just the ids
invert_wide_data <- dplyr::select(invert_wide,-c(year,month,date,site,transect,survey))
invert_wide_data[is.na(invert_wide_data)] <- 0
invert_wide_ids <- dplyr::select(invert_wide,c(year,month,date,site,transect,survey))
# calculate diversity
simpson.div <- diversity(invert_wide_data,index="simpson")
shannon.div <- diversity(invert_wide_data)
  
order.diversity <- invert_wide_ids %>%
  mutate(order.div.simpson = simpson.div, order.div.shannon = shannon.div)
```
<br>

Urchin abundance
```{r}
urchins <- invert %>%
  filter(taxon_family == "Strongylocentrotidae") %>%
  dplyr::select(year, month, date,site,transect, survey,quad,side,taxon_order, sp_code,scientific_name,group,size,count,area) %>%
  filter(year >= 2001) %>%
  filter(!(site == "IVEE" & transect %in% c(3,5,6,7,8))) %>%
  group_by(year,month,date,site,transect,survey,scientific_name) %>%
  summarise(abund = sum(count)) %>%
  pivot_wider(names_from=scientific_name,values_from=abund) %>%
  rename(franciscanus.abundance=`Mesocentrotus franciscanus`,purpuratus.abundance=`Strongylocentrotus purpuratus`) %>%
  mutate(urchin.abundance=sum(franciscanus.abundance, purpuratus.abundance))
```
<br>

Combine variables and write out. 
```{r}
invert_out <- left_join(sp.richness,abundance, by=c("year","month","date","site","transect","survey")) %>%
  left_join(sp.diversity, by=c("year","month","date","site","transect","survey")) %>%
  left_join(order.richness, by=c("year","month","date","site","transect","survey")) %>%
  left_join(order.diversity, by=c("year","month","date","site","transect","survey")) %>%
  left_join(urchins, by=c("year","month","date","site","transect","survey"))

write.csv(invert_out, here::here("data","Annual_Invert_RichnessAbundDiversity_2001-2018.csv"), row.names=FALSE)
```
<br>
<br>

---


## Oceanographic data

### Raw Data

Continuous reef bottom temperature, ongoing since 2000.
```{r}
temp_raw <- read.csv(here::here("data","raw","Bottom_temp_all_years_20190806.csv")) %>%
   clean_names() %>%
  remove_empty("rows") %>%
  na_if(-99999)
colnames(temp_raw); dim(temp_raw)
```
<br>

Water chemistry measurements, ongoing since 2003.
```{r}
water_raw <- read.delim(here::here("data","raw","LTER_monthly_bottledata_20190904.txt"),sep=";") %>%
  clean_names() %>%
  remove_empty("rows") %>%
  na_if(-99999) %>%
  na_if(99999)
colnames(water_raw); dim(water_raw)
```
<br>


### Process Data

#### Reef bottom temperature data

First, I'll just convert the dates into date objects so that I can work with them. 
```{r}
tempdat <- temp_raw %>%
  mutate(date_local=date(parse_date_time(date_local,orders=c("ymd","mdy"))),
         year=year(date_local), month=month(date_local)) %>%
  filter(!is.na(date_local))
min(tempdat$date_local,na.rm=TRUE); max(tempdat$date_local,na.rm=TRUE)
```
<br>

I'm going to average the summer and winter temperatures at each site for each year, and find the variation in temperature throughout those seasons. To do so, I want to see if there are any natural breaks in temperature fluctuations.
```{r}
ggplot(data=filter(tempdat, year==2003 | year==2004), aes(x=date_local,y=temp_c, col=site)) +  geom_smooth()
```


Let's say that "summer" is May - Oct, and "winter" is Nov - April.
```{r}
tempdat_means <- tempdat %>%
  mutate(season=ifelse(month > 5 & month < 10, "summer","winter")) %>%
  group_by(site,year,season) %>%
  summarise(mean.temp=mean(temp_c), var.temp=var(temp_c), sd.temp=sd(temp_c))
```
<br>

Write out
```{r}
write.csv(tempdat_means, here::here("data","Seasonal_Mean_Bottom_Temp_2002-2019.csv"),row.names=FALSE)
```
<br>


#### Water chemistry data

*LTER Abstract:* water chemistry measurements taken monthly at these reefs in the nearshore areas of the Santa Barbara Channel, CA, USA: Arroyo Quemado, Bullito, Naples, Arroyo Burro, Mohawk and Carpinteria. ongoing since 2003. Measurements include standard CTD parameters, nutrients, pigments, particulate CN, total dissolved N and P, stable isotopes of C and N (not all parameters are measured at all stations). Sampling began in November 2000. Some stations are sampled only occasionally. During the first 2 years, CTD data were collected with a SBE19 Seacat Profiler and water samples with a pump. Starting in February 2003, a SBE19-Plus with a rosette sampler was used.

I'm interested in total dissolved nitrogen and phosphorus (how nutrient-rich is the water? can also be an indication of run-off, and kelp is known to require nitrogen to grow). I'll also look at chlorophyll (indicates primary productivity), urea (indicates run-off, as a common fertilizer).

```{r}
water <- water_raw %>%
  dplyr::select(station, mon_day_yr, hh_mm, target_depth_m,
               po4_umol_l,no2_no3_umol_l,poc_umol_l,pon_umol_l,urea_umol_l,ammonia_umol_l,
               total_chl_ug_l) %>%
  na_if("BDL") %>%
  filter(station %in% c("BR","AB","CR","MK","NR","QR")) %>%
  mutate(station = recode(station, BR="BULL", AB ="ABUR", CR="CARP", MK="MOHK", NR="NAPL", QR="AQUE"),
         station=as.character(station)) %>%
  filter(!is.na(target_depth_m)) %>%
  # get a daily average
  group_by(station, mon_day_yr,target_depth_m) %>%
  summarise(po4=mean(as.numeric(po4_umol_l), na.rm=TRUE), no2_no3=mean(as.numeric(no2_no3_umol_l), na.rm=TRUE),
            poc=mean(as.numeric(poc_umol_l),na.rm=TRUE),pon=mean(as.numeric(pon_umol_l),na.rm=TRUE),
            urea=mean(as.numeric(urea_umol_l),na.rm=TRUE),ammonia=mean(as.numeric(ammonia_umol_l),na.rm=TRUE),
            tchl=mean(as.numeric(total_chl_ug_l),na.rm=TRUE)) %>%
  mutate(read_date=mdy(mon_day_yr), year=year(read_date))
```
<br>

How does po4 / total chlorophyll vary by site, across the course of a year?
```{r}
ggplot(data=filter(water,year %in% c(2010,2011) & target_depth_m==1), aes(x=read_date,y=po4, col=station)) + geom_line()

ggplot(data=filter(water,year %in% c(2010,2011) & target_depth_m==1), aes(x=read_date,y=tchl, col=station)) + geom_line()
```
<br>

I think I'll take the annual mean and maximum of each nutrient concentration. I'm going to focus on nutrients at the surface, 1-5m target depth, and I'll average across those depths.
```{r}
water_means <- water %>%
  filter(target_depth_m ==1) %>%
  pivot_longer(cols=c(po4,no2_no3,pon,poc,urea,ammonia,tchl), names_to="ctd.var", values_to="value") %>%
  mutate(value=as.numeric(value),
         value=ifelse(is.nan(value),NA,value)) %>%
  group_by(station,year,ctd.var) %>%
  summarise(mean=mean(as.numeric(value),na.rm=TRUE), max=0) %>%
  pivot_wider(names_from=ctd.var,values_from=c(mean,max)) 
water_means <- water_means[,1:8]

water_maxes <- water %>%
  filter(target_depth_m ==1) %>%
   pivot_longer(cols=c(po4,no2_no3,pon,poc,urea,ammonia,tchl), names_to="ctd.var", values_to="value") %>%
   mutate(value=as.numeric(value),
         value=ifelse(is.nan(value),NA,value)) %>%
  filter(!is.na(value)) %>%
  group_by(station,year,ctd.var) %>%
  summarise(max=max(as.numeric(value),na.rm=TRUE),mean=0) %>%
  pivot_wider(names_from=ctd.var,values_from=c(max,mean))
water_out <- water_maxes[,1:8] %>%
  left_join(water_means, by=c("station","year")) %>%
  arrange(year)
head(water_out)
```
<br>

write out.
```{r}
write.csv(water_out, here::here("data","Annual_CTD_WaterChemistry_MeanMax_2000-2018.csv"),row.names=FALSE)
```
<br>
