==> scripts/03_data_cleaning_biomass.Rmd <==
---
title: "Pre-process Biomass Data"
author: "Mary Fisher"
date: "February 28, 2020"
output: 
  html_document:
    toc: yes
    toc_float: yes
---
<br>
<br>

We found minimal correlation between giant kelp size and abundance, and the fish and invertebrate metrics of abundance / richness / diversity. This finding is contrary to existing scientific research and knowledge of kelp forest ecosystems, and may have been a function of they way that I cleaned and summarized the SBC LTER data sets. I'm going to look at the biomass data provided by the SBC LTER, which has been summarized at the transect (replicate / site) level. My hope is that scientists with the SBC LTER, who have a more extensive understanding of this data set, may have cleaned and summarized their survey data using a different - and more appropriate - approach than I did.

Biomass is a measure of the mass of a given biological community. It can be used in place of raw count data to represent abundance, diversity, and community composition. 

<br>

```{r setup, include=FALSE}
rm(list=ls())

knitr::opts_chunk$set(echo = FALSE)
library(tidyverse)
library(janitor)
library(here)
library(vegan)
library(lubridate)
library(corrplot)
library(MASS)
```

## Raw Data

*SBCLTER Abstract*: These data are annual estimates of biomass of approximately 225 taxa of reef algae, invertebrates and fish in permanent transects at 11 kelp forest sites in the Santa Barbara Channel (2-8 transects per site). Abundance is measured annually (as percent cover or density, by size) and converted to biomass (i.e., wet mass, dry mass, decalcified dry mass, ash free dry mass) using published taxon-specific algorithms. Data collection began in summer 2000 and continues annually in summer to provide information on community structure, population dynamics and species change. The time period of data collection varied among the 11 kelp forest sites. Sampling at BULL, CARP, and NAPL began in 2000, sampling at the other 6 mainland sites (AHND, AQUE, IVEE, GOLB, ABUR, MOHK) began in 2001 (transects 3, 5, 6, 7, 8 at IVEE were added in 2011). Data collection at the two Santa Cruz Island sites (SCTW and SCDI) began in 2004. See Methods for more information.   See Methods for more information.   

```{r}
biomass <- read.csv(here::here("data","raw","Annual_All_Species_Biomass_at_transect_20200108.csv")) %>%
  clean_names() %>%
  na_if(99999) %>%
  na_if(-99999)
dim(biomass)
colnames(biomass)
```
<br>

How much missing data is there across the different biomass calculations per group? 
```{r}
biomass %>% group_by(group) %>% summarise(wm=sum(!is.na(wm_gm2))/length(wm_gm2),
                                          dry=sum(!is.na(dry_gm2))/length(dry_gm2),
                                          sf=sum(!is.na(sfdm))/length(sfdm),
                                          ash=sum(!is.na(afdm))/length(afdm))
```
Wet mass and dry mass calculations have the least missing data. I'll stick with the dry mass calculations for now. 

<br>

## Transect-level Biomass

### Process data

Total and mean biomass per group / site / transect. 
```{r}
total_biomass <- biomass %>%
  filter(group != "ALGAE") %>%
  group_by(year,month,site,transect,group) %>%
  summarise(total_biomass = sum(dry_gm2,na.rm=TRUE), mean_biomass=mean(dry_gm2, na.rm=TRUE)) %>%
  pivot_longer(cols=c(total_biomass,mean_biomass)) %>%
  pivot_wider(names_from=c(group,name))

kelp_biomass <- biomass %>%
  filter(scientific_name=="Macrocystis pyrifera") %>%
  group_by(year,month,site,transect) %>%
  summarise(kelp_total_biomass = sum(dry_gm2,na.rm=TRUE), kelp_mean_biomass=mean(dry_gm2, na.rm=TRUE))

nonkelp_algae <- biomass %>%
  filter(group == "ALGAE" & scientific_name != "Macrocystis pyrifera") %>%
  group_by(year,month,site,transect) %>%
  summarise(algae_total_biomass = sum(dry_gm2,na.rm=TRUE), algae_mean_biomass=mean(dry_gm2, na.rm=TRUE))
```
<br>

```{r}
biomass_summary <- full_join(total_biomass,nonkelp_algae,by=c("year","month","site","transect")) %>%
  full_join(kelp_biomass, by=c("year","month","site","transect")) %>%
  clean_names()
head(biomass_summary)
```
<br>

### Relationship between Variables

```{r}
corrplot(cor(biomass_summary[5:12], use="complete.obs"), type="upper")
```
<br>

Fish & Kelp
```{r}
plot(biomass_summary[,c(5,6,11,12)])
```
<br>

Invertebrates & Kelp
```{r}
plot(biomass_summary[,c(7,8,11,12)])
```
<br>

Algae & Kelp
```{r}
plot(biomass_summary[,10:12])
```
<br>

Fish & Invertebrates
```{r}
plot(biomass_summary[,5:8])
```
<br>


## Site-level Biomass

### Process Data

What if we look at the site level, instead of at the level of transects? This would account for spatial heterogeneity in kelp, fish and invertebrate biomass across transects, assuming that it is the site-level biomass that actually matters. 

```{r}
total_biomass <- biomass %>%
  filter(group != "ALGAE") %>%
  group_by(year,month,site,group) %>%
  summarise(total_biomass = sum(dry_gm2,na.rm=TRUE), mean_biomass=mean(dry_gm2, na.rm=TRUE)) %>%
  pivot_longer(cols=c(total_biomass,mean_biomass)) %>%
  pivot_wider(names_from=c(group,name))

kelp_biomass <- biomass %>%
  filter(scientific_name=="Macrocystis pyrifera") %>%
  group_by(year,month,site) %>%
  summarise(kelp_total_biomass = sum(dry_gm2,na.rm=TRUE), kelp_mean_biomass=mean(dry_gm2, na.rm=TRUE))

nonkelp_algae <- biomass %>%
  filter(group == "ALGAE" & scientific_name != "Macrocystis pyrifera") %>%
  group_by(year,month,site) %>%
  summarise(algae_total_biomass = sum(dry_gm2,na.rm=TRUE), algae_mean_biomass=mean(dry_gm2, na.rm=TRUE))

biomass_site_summary <- full_join(total_biomass,nonkelp_algae,by=c("year","month","site")) %>%
  full_join(kelp_biomass, by=c("year","month","site")) %>%
  clean_names()
head(biomass_site_summary)
```
<br>

I'll also calculate the Simpson diversity index using biomass. 
```{r}
# fish
div_fish_dat <- biomass %>%
  filter(group=="FISH" & !is.na(taxon_order)) %>%
  group_by(year,month,site,taxon_order) %>%
  summarise(total_biomass=sum(dry_gm2,na.rm=TRUE)) %>%
  pivot_wider(names_from=taxon_order,values_from=total_biomass)
div_fish <- biomass %>%
  filter(group=="FISH" & !is.na(taxon_order)) %>%
  group_by(year,month,site,taxon_order) %>%
  summarise(total_biomass=sum(dry_gm2,na.rm=TRUE)) %>%
  pivot_wider(names_from=taxon_order,values_from=total_biomass) %>%
  dplyr::select(year,month,site)
div_fish[,"fish_diversity"] <- diversity(div_fish_dat[,c(-1,-2,-3)],index="invsimpson")

# invertebrates
div_invert_dat <- biomass %>%
  filter(group=="INVERT" & !is.na(taxon_order)) %>%
  group_by(year,month,site,taxon_order) %>%
  summarise(total_biomass=sum(dry_gm2,na.rm=TRUE)) %>%
  pivot_wider(names_from=taxon_order,values_from=total_biomass)
div_invert <- biomass %>%
  filter(group=="INVERT" & !is.na(taxon_order)) %>%
  group_by(year,month,site,taxon_order) %>%
  summarise(total_biomass=sum(dry_gm2,na.rm=TRUE)) %>%
  pivot_wider(names_from=taxon_order,values_from=total_biomass) %>%
  dplyr::select(year,month,site)
div_invert[,"invert_diversity"] <- diversity(div_invert_dat[,c(-1,-2,-3)],index="invsimpson")

# non-kelp algae
div_algae_dat <- biomass %>%
  filter(group=="ALGAE" & scientific_name != "Macrocystis pyrifera" & !is.na(taxon_order)) %>%
  group_by(year,month,site,taxon_order) %>%
  summarise(total_biomass=sum(dry_gm2,na.rm=TRUE)) %>%
  pivot_wider(names_from=taxon_order,values_from=total_biomass)
div_algae <- biomass %>%
  filter(group=="ALGAE" & scientific_name != "Macrocystis pyrifera" & !is.na(taxon_order)) %>%
  group_by(year,month,site,taxon_order) %>%
  summarise(total_biomass=sum(dry_gm2,na.rm=TRUE)) %>%
  pivot_wider(names_from=taxon_order,values_from=total_biomass) %>%
  dplyr::select(year,month,site)
div_algae[,"algae_diversity"] <- diversity(div_algae_dat[,c(-1,-2,-3)],index="invsimpson")
div_algae$algae_diversity[div_algae$algae_diversity==Inf] <- 0

biomass_div <- full_join(div_fish, div_invert, by=c("year","month","site")) %>%
  full_join(div_algae,by=c("year","month","site"))
```
```{r}
biomass_site_summary <- biomass_site_summary %>%
  full_join(biomass_div, by=c("year","month","site"))
head(biomass_site_summary)
```
<br>

### Relationship between variables

```{r}
corrplot(cor(biomass_site_summary[4:11], use="complete.obs"), type="upper")
```
<br>

There seem to be some relationships between invertebrate - algae biomass, and algae - kelp biomass. 

```{r}
corrplot(cor(biomass_site_summary[,c(10:14)], use="complete.obs"), type="upper")
```

```{r}
plot(biomass_site_summary[4:11])
plot(biomass_site_summary[,c(10:14)])
```

### Write out

First, all of the biomass metrics at the site-level.
```{r}
colnames(biomass_site_summary)
write.csv(biomass_site_summary, here::here("data","all_biomass_metrics_bysite.csv"), row.names=FALSE)
```
<br>

Then, a thinned-out data set that contains only total biomass and shannon diversity
```{r}
biomass_thin <- dplyr::select(biomass_site_summary,
                              year,month,site,
                              fish_total_biomass,fish_diversity,invert_total_biomass,invert_diversity,algae_total_biomass,algae_diversity,
                              kelp_total_biomass)
colnames(biomass_thin)

write.csv(biomass_thin, here::here("data","biomass_prediction_data_bysite.csv"),row.names=FALSE)
```
<br>

Finally, read in the water chemistry variables and include that in the data set.
```{r}
temp <- read.csv(here::here("data","Seasonal_Mean_Bottom_Temp_2002-2019.csv"))  %>%
  dplyr::select(year,site,season,mean.temp,var.temp) %>%
  rename(mean_temp=mean.temp,var_temp=var.temp) %>%
  pivot_longer(cols=c(mean_temp,var_temp)) %>%
  pivot_wider(names_from=c(season,name), values_from=value)

chem <- read.csv(here::here("data","Annual_CTD_WaterChemistry_MeanMax_2000-2018.csv"))

biomass_waterchem <- left_join(biomass_thin, temp, by=c("site","year")) %>%
  filter(!is.na(summer_mean_temp)) %>%
  left_join(chem,by=c("site"="station","year")) %>%
  filter(site %in% chem$station) %>%
  filter(year != 2019 & year != 2001)

biomass_waterchem_complete <- biomass_waterchem %>%
  filter(site %in% chem$station) %>%
  filter(!is.na(mean_po4))
colnames(biomass_waterchem_complete)
```
<br>

```{r}
write.csv(biomass_waterchem, here::here("data","biomass_prediction_data_waterchem_bysite.csv"),row.names=FALSE)
write.csv(biomass_waterchem_complete, here::here("data","biomass_prediction_data_waterchem_bysite_complete.csv"),row.names=FALSE)
```
<br>
