==> scripts/04_explore_predictors.Rmd <==
---
title: "Explore Variables"
author: "Mary Fisher"
date: "March 4, 2020"
output: 
  html_document:
    toc: yes
    toc_float: yes
---

Check for normality and multicollinearity of predictor variables. 



```{r setup, include=FALSE}
rm(list=ls())

knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(janitor)
library(here)
library(vegan)
library(lubridate)
library(corrplot)
library(MASS)

theme_set(theme_bw())
```
<br>

Read in data
```{r}
mydat1 <- read.csv(here::here("data","kelp_prediction_data_complete.csv")) %>%
  clean_names()
colnames(mydat1)

mydat2 <- read.csv(here::here("data","biomass_prediction_data_waterchem_bysite_complete.csv")) %>%
  clean_names()
colnames(mydat2)
```
<br>


# Question 1: Kelp Growth & Size

## Distributions

### Predictor: Herbivory

*These variables are zero-inflated. a transformation is unlikely to work completely. Instead, we would have two predictor variables - are there urchins (categorical, 1/0) and if there are, how abundant are they?*

```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=colnames(mydat1)[c(10:12)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```
<br>

Square Root of Variable
```{r}
ggplot(data=plotdat, aes(x=sqrt(value))) + 
  geom_density() + 
  facet_wrap(~variable,ncol=2,scales="free")
shapiro.test(sqrt(as.numeric(mydat1$urchin_abundance)))
```
<br>

Just looking at the values > 0
```{r}
ggplot(data=filter(mydat1,urchin_abundance > 0), aes(x=urchin_abundance)) + 
  geom_histogram()
```
<br>

Square Root of Variable
```{r}
ggplot(data=filter(mydat1,urchin_abundance > 0), aes(x=sqrt(urchin_abundance))) + 
  geom_histogram()
shapiro.test(sqrt(as.numeric(filter(mydat1, urchin_abundance>0)$urchin_abundance)))
```
<br>

### Predictor: Water Chemistry

#### Max Values
```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=colnames(mydat1)[c(13:18)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```
<br>

Log of max variables
```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=c(max_pon,max_poc,max_tchl), names_to="variable")
ggplot(data=plotdat, aes(x=log(value))) + 
  geom_density() + 
  facet_wrap(~variable,ncol=2,scales="free")
mydat1 %>%
  dplyr::select(max_pon,max_poc,max_tchl) %>%
  map(function(x) shapiro.test(as.numeric(log(x))))
```

Max variables to the third power
```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=c(max_ammonia,max_no2_no3,max_po4), names_to="variable")
ggplot(data=plotdat, aes(x=value^3)) + 
  geom_density() + 
  facet_wrap(~variable,ncol=2,scales="free")

mydat1 %>%
  dplyr::select(max_ammonia,max_no2_no3,max_po4) %>%
  map(function(x) shapiro.test(as.numeric((x^3))))
```
<br>

#### Mean values
```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=colnames(mydat1)[c(19:24)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```
<br>

Log of mean variables. Mostly works!
```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=c(mean_pon,mean_poc,mean_tchl), names_to="variable")
ggplot(data=plotdat, aes(x=log(value))) + 
  geom_density() + 
  facet_wrap(~variable,ncol=2,scales="free")

mydat1 %>%
  dplyr::select(mean_pon,mean_poc,mean_tchl) %>%
  map(function(x) shapiro.test(as.numeric(log(x))))
```



### Predictor: Water Temperature

```{r}
plotdat <- mydat1 %>%
  pivot_longer(cols=colnames(mydat1)[c(25:28)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")

mydat1 %>%
  dplyr::select(colnames(mydat1)[c(25:28)]) %>%
  map(function(x) shapiro.test(as.numeric(x)))
```

Log-transform summer temperatures
```{r}
mydat1 %>%
  dplyr::select(colnames(mydat1)[c(25:26)]) %>%
  map(function(x) shapiro.test(as.numeric(log(x))))
```
<br>

Square root of winter temperatures
```{r}
mydat1 %>%
  dplyr::select(colnames(mydat1)[c(27:28)]) %>%
  map(function(x) shapiro.test(as.numeric(sqrt(x))))
```
<br>


## Multicollinearity

### Max values

```{r}
corrplot(cor(mydat1[,c(10:18,25:28)],use="complete.obs"), type="upper")
```
<br>

### Mean values
```{r}
corrplot(cor(mydat1[,c(10:12,19:28)], use="complete.obs"), type="upper")
```
<br>


# Question 2: Biodiversity

## Distributions

### Predictor: Water Temperature
```{r}
plotdat <- dplyr::select(mydat2, -year, -month) %>%
  pivot_longer(cols=colnames(mydat2)[c(11:14)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```

*see transformations above*
<br>

### Predictor: Water Chemistry
```{r}
plotdat <- dplyr::select(mydat2, -year, -month) %>%
  pivot_longer(cols=colnames(mydat2)[c(15:20)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```
```{r}
plotdat <- dplyr::select(mydat2, -year, -month) %>%
  pivot_longer(cols=colnames(mydat2)[c(21:26)], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```
<br>

*see transformations above*

### Predictor: Kelp

```{r}
plotdat <- dplyr::select(mydat2, -year, -month) %>%
  pivot_longer(cols=colnames(mydat2)[10], names_to="variable")
ggplot(data=plotdat, aes(x=value)) + 
  geom_histogram() + 
  facet_wrap(~variable,ncol=2,scales="free")
```
<br>

Cubed root of variable
```{r}
ggplot(data=plotdat, aes(x=value^(1/3))) + 
  geom_density() + 
  facet_wrap(~variable,ncol=2,scales="free")
shapiro.test((mydat2$kelp_total_biomass)^(1/3))
```

## Multicollinearity

### Max values

```{r}
corrplot(cor(mydat2[,c(10:20)],use="complete.obs"), type="upper")
```
<br>

### Mean values
```{r}
corrplot(cor(mydat2[,c(10:14,21:26)],use="complete.obs"), type="upper")
```
