==> scripts/05_linear_models_q1.Rmd <==
---
title: "Variable Selection Question 2"
author: "Mary Fisher"
date: "February 28, 2020"
output: html_document
---

#### Research Question: Can we predict giant kelp size and abundance from oceanographic conditions and herbivore abundance (predictor)?

**Response variables:** Giant kelp mean abundance, max abundance; Giant kelp mean size, maximum size. 

**Predictor variables: **

1. Urchin abundance: Sea urchins are herbivores, and in large numbers (also called "herds") can decimate kelp forests. We have three measures of urchin abundance - red urchin abundance, purple urchin abundance, and urchin abundance across all species.

2. Water chemistry: Like any algae / plant, kelp requires nutrients to grow. We have measurements for the following nutrients: nitrogen, ammonia, urea, and phosphorus. We also have chlorophyl concentrations, which is a general index of primary productivity.

3. Water temperature: All species have a thermal range, or a range of temperatures, at which they grow best; the degree of variability in temperature can also have an impact on growth. With kelp forest loss from recent marine heatwaves and long-term warming, there has been much interest in whether kelp are impacted by an absolute temperature threshold, or by the degree of variability in temperature. We have mean summer and mean winter temperatures, and variation in summer and mean winter temperatures. 

4. Giant kelp mean / max abundance, mean / max size at t-1: This data set is a time series, so the giant kelp response variable at a given site for year `t` should be dependent on giant kelp condition in year `t-1`.



```{r setup, include=FALSE}
rm(list=ls())

knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(here)
library(vegan)
library(lubridate)
library(corrplot)
library(MASS)
```
<br>

Read in data
```{r}
mydat <- read.csv(here::here("data","kelp_prediction_data.csv")) %>%
  clean_names()
head(mydat)
```
<br>

Add in the giant kelp response variables for `t-1`.
```{r}
# need kelp info from 2001
kelp_dat <- read.csv(here::here("data","Annual_Kelp_MeanSize_MeanAbund_2001-2018.csv")) %>%
  clean_names() %>%
  filter(site %in% mydat$site & year == 2001) %>%
  dplyr::select(-scientific_name,-group, -month) %>%
  rename("prev_mean_abund" = mapy_mean_abund,"prev_max_abund"=mapy_max_abund,
         "prev_mean_size"=mapy_mean_size,"prev_max_size"=mapy_max_size) %>%
  mutate(year = year + 1)
kelp_prev <- mydat %>% dplyr::select(year,month,site,transect,
                                     mapy_mean_abund,mapy_max_abund,mapy_mean_size,mapy_max_size) %>%
  rename("prev_mean_abund" = mapy_mean_abund,"prev_max_abund"=mapy_max_abund,
         "prev_mean_size"=mapy_mean_size,"prev_max_size"=mapy_max_size) %>%
  mutate(year = year + 1) %>%
  dplyr::select(-month) %>%
  rbind(kelp_dat)
mydat <- left_join(mydat,kelp_prev,by=c("year","site","transect"))

message("\nThere are ", sum(is.na(mydat$prev_mean_abund)), " missing measurements for the previous year's kelp abundance.\n")
head(mydat %>% dplyr::select(year,site,transect,mapy_mean_abund,prev_mean_abund,prev_max_abund,prev_mean_size,prev_max_size))
```
<br>


## Multicolinearity / Predictor Selection

```{r}
corrplot(cor(mydat[,c(9:17,29,31)],use="complete.obs"), type="upper")
corrplot(cor(mydat[,c(18:28,30)], use="complete.obs"), type="upper")
```
<br>

We have three variables that quantify urchin abundance - we'll just work with total abundance (`urchin_abundance`). 




Nitrogen and phosphorus are correlated, so we'll choose to look at nitrogen (`no2_no3`) because nitrogen is a key nutrient required for kelp growth. We also focused on mean seasonal temperatures, rather than the variability in seasonal temperature. 

```{r}
lmdat <- mydat %>%
  dplyr::select(-purpuratus_abundance,-franciscanus_abundance, -max_po4, -mean_po4)
```
<br>


## Model Selection & Diagnostics

### Size of Giant Kelp

#### Boxcox

```{r}
library(car)
dat <- lmdat %>% filter(mapy_mean_fronds > 0)
fit1 <- lm(mapy_mean_fronds~.,data=dat, na.action = na.exclude)
boxcox(fit1)
summary(powerTransform(fit1))
findlambda <- boxcox(mapy_mean_fronds~., data = dat,lambda = seq(-1,1,length = 20))
findlambda$x[which.max(findlambda$y)]

hist(lmdat$mapy_mean_fronds)
lmdat$mapy_mean_fronds <- (lmdat$mapy_mean_fronds^0.5 -1)/0.5
hist(lmdat$mapy_mean_fronds)
```	

#### Fit model
```{r}
n <- dim(lmdat)[1]
fit.a <- lm(mapy_mean_fronds ~ urchin_abundance+site+
              mean_no2_no3+mean_tchl+
              summer_mean_temp+winter_mean_temp+
              mean_no2_no3*summer_mean_temp+mean_tchl*summer_mean_temp +
              site*summer_mean_temp,data=lmdat)
scp <- list(lower = ~ 1,
           upper = ~ urchin_abundance+site+
              mean_no2_no3+mean_tchl+
              summer_mean_temp+winter_mean_temp+
              mean_no2_no3*summer_mean_temp+mean_tchl*summer_mean_temp+site*summer_mean_temp)

fit.f <- step(fit.a, scope = scp, direction = "backward",  k = log(n))
summary(fit.f)
```
<br>

```{r}
n <- dim(lmdat)[1]
fit.e <- lm(mapy_mean_fronds ~ 1,data=lmdat)
fit.a <- lm(mapy_mean_fronds ~ urchin_abundance+site+
              mean_no2_no3+mean_tchl+
              summer_mean_temp+winter_mean_temp+
              mean_no2_no3*summer_mean_temp+mean_tchl*summer_mean_temp+
              site*summer_mean_temp,data=lmdat)
scp <- list(lower = ~ 1,
           upper = ~ urchin_abundance+site+
              mean_no2_no3+mean_tchl+
              summer_mean_temp+winter_mean_temp+
              mean_no2_no3*summer_mean_temp+mean_tchl*summer_mean_temp)

fit.f <- step(fit.e, scope = scp, direction = "both",  k = log(n))
```
<br>

Final model: `mapy_mean_size ~ site + urchin_abundance + summer_mean_temp + mean_tchl + mean_no2_no3 + summer_mean_temp:mean_no2_no3`
```{r}
summary(fit.f)
```
<br>
How does RSE / R squared compare to full model
```{r}
summary(fit.a)
```
<br>

#### Model Diagnostics

It looks like this model violates the normality assumption, but there don't appear to be any outliers. The TA plot looks a little weird. 
```{r}
par(mfrow=c(2,2))
plot(fit.f)
```
<br>

Here is the right tail on the residuals:
```{r}
hist(fit.f$residuals)
```
<br>

    


### Abundance of Giant Kelp

#### Boxcox

```{r}
library(car)
dat <- lmdat %>% filter(mapy_mean_density > 0)
fit1 <- lm(mapy_mean_density~.,data=dat, na.action = na.exclude)
boxcox(fit1)
summary(powerTransform(fit1))
findlambda <- boxcox(mapy_mean_density~., data = dat,lambda = seq(-1,1,length = 20))
findlambda$x[which.max(findlambda$y)]

```	

#### Fit Model
```{r}
n <- dim(lmdat)[1]
fit.abund.e <- lm(mapy_mean_density ~ 1,data=lmdat)
fit.abund.a <- lm(mapy_mean_density ~ urchin_abundance+site+
              mean_no2_no3+mean_tchl+
              summer_mean_temp+winter_mean_temp+
              mean_no2_no3*summer_mean_temp+mean_tchl*summer_mean_temp+
              site*summer_mean_temp,data=lmdat)
scp <- list(lower = ~ 1,
           upper = ~ urchin_abundance+site+
              mean_no2_no3+mean_tchl+
              summer_mean_temp+winter_mean_temp+
              mean_no2_no3*summer_mean_temp+mean_tchl*summer_mean_temp)

fit.abund.f <- step(fit.abund.e, scope = scp, direction = "forward",  k = log(n))
```

```{r}
fit.abund.f <- step(fit.abund.a, direction = "backward",  k = log(n))
```
<br>

```{r}
summary(fit.abund.f)
summary(fit.abund.a)
```
<br>

#### Model Diagnostics

A bunch of things look weird here: the TA plot shows a fan shape, suggesting that our estimates are less precise at higher values. The model also violates the assumption of normality. 
```{r}
par(mfrow=c(2,2))
plot(fit.abund.f)
```
<br>

Here is the right tail on the residuals:
```{r}
hist(fit.abund.f$residuals)
```
