==> scripts/05_linear_models_timeseries_q1.Rmd <==
---
title: "Question 1: with t-1 variable"
author: "Mary Fisher"
date: "February 28; Updated March 13"
output: 
  html_document:
    toc: yes
    toc_float: yes
---

#### Research Question: Which oceanographic and herbivory conditions are important for predicting giant kelp size and abundance?

**Response variables:** Giant kelp mean abundance, max abundance; Giant kelp mean size, maximum size. 

**Predictor variables: **

1. Urchin abundance: Sea urchins are herbivores, and in large numbers (also called "herds") can decimate kelp forests. We have three measures of urchin abundance - red urchin abundance, purple urchin abundance, and urchin abundance across all species.

2. Water chemistry: Like any algae / plant, kelp requires nutrients to grow. We have measurements for the following nutrients: nitrogen, ammonia, urea, and phosphorus. We also have chlorophyl concentrations, which is a general index of primary productivity.

3. Water temperature: All species have a thermal range, or a range of temperatures, at which they grow best; the degree of variability in temperature can also have an impact on growth. With kelp forest loss from recent marine heatwaves and long-term warming, there has been much interest in whether kelp are impacted by an absolute temperature threshold, or by the degree of variability in temperature. We have mean summer and mean winter temperatures, and variation in summer and mean winter temperatures. 

4. Giant kelp mean / max abundance, mean / max size at t-1: This data set is a time series, so the giant kelp response variable at a given site for year `t` should be dependent on giant kelp condition in year `t-1`.


**Modeling approach:** OLS incorporating previous size / abundance.



```{r setup, include=FALSE}
rm(list=ls())

knitr::opts_chunk$set(echo = FALSE)
library(tidyverse)
library(janitor)
library(here)
library(vegan)
library(lubridate)
library(corrplot)
library(MASS)
library(performance)

```
<br>

# Prepare Data

Read in data
```{r echo=TRUE}
mydat <- read.csv(here::here("data","kelp_prediction_data_complete.csv")) %>%
  clean_names()
head(mydat)
```
<br>

Add in the giant kelp response variables for `t-1`.
```{r}
# need kelp info from 2001
kelp_dat <- read.csv(here::here("data","Annual_Kelp_MeanSize_MeanAbund_2001-2018.csv")) %>%
  clean_names() %>%
  filter(site %in% mydat$site & year == 2001) %>%
  dplyr::select(-month) %>%
  rename("prev_mean_density" = mapy_mean_density,
         "prev_mean_count"=mapy_mean_count,
         "prev_mean_fronds"=mapy_mean_fronds,
         "prev_max_fronds"=mapy_max_fronds,
         "prev_mean_hld"=mapy_mean_hld,"prev_max_hld"=mapy_max_hld) %>%
  mutate(year = year + 1)
kelp_prev <- read.csv(here::here("data","kelp_prediction_data.csv")) %>%
  clean_names() %>%
  dplyr::select(year,month,site,mapy_mean_density,mapy_mean_count,mapy_mean_fronds,mapy_max_fronds,mapy_mean_hld,mapy_max_hld) %>%
  rename("prev_mean_density" = mapy_mean_density,
         "prev_mean_count"=mapy_mean_count,
         "prev_mean_fronds"=mapy_mean_fronds,
         "prev_max_fronds"=mapy_max_fronds,
         "prev_mean_hld"=mapy_mean_hld,"prev_max_hld"=mapy_max_hld) %>%
  mutate(year = year + 1) %>%
  dplyr::select(-month) %>%
  rbind(kelp_dat)
mydat <- left_join(mydat,kelp_prev,by=c("year","site"))

message("\nThere are ", sum(is.na(mydat$prev_mean_density)), " missing measurements for the previous year's kelp abundance.\n")

colnames(mydat)
```
<br>

## Predictor Variables 

### Herbivory

We have three variables that quantify urchin abundance - we'll just work with total abundance (`urchin_abundance`). 

This variable is zero-inflated, a common problem with ecological count data. So I'm going to split it into two predictor variables - a categorical presence / absence variable, and then an abundance variable.

First, get rid of the species-specific counts
```{r echo=TRUE}
lmdat <- dplyr::select(mydat, -purpuratus_abundance,-franciscanus_abundance)
```
<br>

Then grab the total abundance data and split it into the two new variables. Does the square root transformation of the urchin abundance variable look normal?
```{r echo=TRUE}
lmdat <- lmdat %>%
  mutate(urchin_presence = ifelse(urchin_abundance > 0, 1, 0),
         urchin_abundance = ifelse(urchin_abundance==0,0,sqrt(urchin_abundance)))
```
```{r}
ggplot(data=filter(lmdat, urchin_abundance > 0),aes(x=urchin_abundance)) +
  geom_histogram() + theme_bw()

shapiro.test(lmdat$urchin_abundance)
```
<br>

### Water Chemistry

The maximum values of water chemistry estimates were very difficult to force into a normal distribution. So I'm only going to work with the mean values.

```{r echo=TRUE}
lmdat <- lmdat %>%
  dplyr::select(-max_ammonia,-max_no2_no3,-max_po4,-max_poc,-max_pon,-max_tchl)
colnames(lmdat)
```
<br>


### Multicolinearity

Let's ignore the previous kelp size / abundance for now.

```{r}
corrplot(cor(lmdat[,c(10:20)],use="complete.obs"), type="upper")
```
<br>

Summer and winter temperatures are correlated; since kelp is a cold-water species, we'll stick with summer temperature only. I'll rely on variable selection to remove either mean or variance in temperature, which are also correlated with each other.

```{r echo=TRUE}
lmdat <- lmdat %>%
  dplyr::select(-winter_mean_temp,-winter_var_temp)
```
<br>

```{r}
corrplot(cor(lmdat[,c(10:18)],use="complete.obs"), type="upper")
```

Particulate organic carbon (`poc`) is highly correlated with particulate organic nitrogen and mean total chlorophyll. I'll remove particulate organic carbon.

```{r echo=TRUE}
lmdat <- lmdat %>%
  dplyr::select(-mean_poc)
```

```{r}
corrplot(cor(lmdat[,c(10:17)],use="complete.obs"), type="upper")
```

I think that for the remaining correlations, I'll see what happens in stepwise model selection. These are the final variables:
```{r}
colnames(lmdat)
```
<br>

### Transform variables
According to the script [04_explore_predictors](https://github.com/mfisher5/504_BiodivSB/tree/master/scripts), many of the predictor variables are not normally distributed. 

```{r echo=TRUE}
lmdat.t <- lmdat %>%
  mutate(summer_var_temp = log(summer_var_temp),
         mean_pon= log(mean_pon),
         mean_tchl= log(mean_tchl))
```
<br>
```{r}
par(mfrow=c(2,2))
hist(lmdat.t$urchin_abundance, main="")
hist(lmdat.t$mean_ammonia, main="")
hist(lmdat.t$mean_no2_no3, main="")
hist(lmdat.t$mean_po4, main="")
hist(lmdat.t$mean_pon, main="")
hist(lmdat.t$mean_tchl, main="")
hist(lmdat.t$summer_mean_temp, main="")
hist(lmdat.t$summer_var_temp, main="")
```
<br>
<br>


## Response Variables

Mean holdfast diameter is normal (yay!).

Mean density and mean number of fronds need to be log-transformed. I think that means that I can just run a generalized linear model using the `family=Gamma(link='log')` argument. 

But I do need to run the variable transformation for the previous mean density / fronds
```{r echo=TRUE}
lmdat.t <- lmdat.t %>%
  mutate(prev_mean_density = log(prev_mean_density),
         prev_mean_count = log(prev_mean_count),
         prev_mean_fronds = log(prev_mean_fronds))
```
```{r}
par(mfrow=c(1,3))
hist(lmdat.t$prev_mean_density,main="")
hist(lmdat.t$prev_mean_count,main="")
hist(lmdat.t$prev_mean_fronds,main="")
```
<br>

Maximum number of fronds and holdfast diameter need a square root transformation. For the response variable, this means that I can just run a generalized linear model using `family=Gamma(link='sqrt')`. 

But I do need to transform the previous fronds and holdfast diameter.
```{r echo=TRUE}
lmdat.t <- lmdat.t %>%
  mutate(prev_max_fronds = sqrt(prev_max_fronds),
         prev_max_hld = sqrt(prev_max_hld))
```
```{r}
par(mfrow=c(1,2))
hist(lmdat.t$prev_max_fronds,main="")
hist(lmdat.t$prev_max_hld,main="")
```
<br>

So now I have the following variables:
```{r}
colnames(lmdat.t)
```
<br>
<br>
<br>


---

# Model Selection


## Kelp Abundance

*Response:* log(`mapy_mean_density`)

Interactions: site by previous mean density

Since there are some missing data in the kelp size categories, I'm going to select only the variables we are interested in at the moment

```{r}
abund_dat <- dplyr::select(lmdat.t, mapy_mean_density,site, urchin_presence, urchin_abundance, prev_mean_density, mean_ammonia, mean_no2_no3, mean_po4, mean_tchl, mean_pon, summer_mean_temp, summer_var_temp)

abund_dat$urchin_presence <- as.factor(abund_dat$urchin_presence)
abund_dat$site <- as.factor(abund_dat$site)
```
<br>


### LM

Model Type: OLS, Log transformation of response

#### Model Selection: BIC

Backward model selection

```{r}
n <- dim(abund_dat)[1]
fit.a <- lm(log(mapy_mean_density) ~ site + urchin_presence + urchin_abundance + 
              prev_mean_density + site:prev_mean_density +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp, data=abund_dat)

fit.fb <- step(fit.a, direction = "backward",  k = log(n))
```
<br>

Forward Model Selection 

```{r}
fit.e <- lm(log(mapy_mean_density) ~ 1, data=abund_dat)

scp <- list(lower = ~ 1,
           upper = ~ site + urchin_presence + urchin_abundance + 
              prev_mean_density + site:prev_mean_density +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp)

fit.ff <- step(fit.e, scope=scp, direction = "forward",  k = log(n))
```
<br>

Forward and backward selection produces slightly different models. The R-squared is better when including the variable `urchin presence`, and that variable is significant.
```{r}
summary(fit.ff); summary(fit.fb)
```
<br>

Model Diagnostics
```{r}
plot(fit.ff)
plot(fit.ff,which=4)
```
<br>

Even with the log-transformation of the response variable, there is still some non-normality of the residuals.

<br>
<br>

#### Model Selection: AIC

Backward model selection

```{r}
fit.fb <- step(fit.a, direction = "backward",  k = 2)
```
<br>

Forward Model Selection 

```{r}
fit.ff <- step(fit.e, scope=scp, direction = "forward",  k = 2)
```
<br>

Forward and backward selection produces *almost* the same model. Since `urchin presence` doesn't really affect the R-squared or the MSE, I'm going to choose the simpler model from the backward selection.
```{r}
summary(fit.ff); summary(fit.fb)
```
<br>

Model Diagnostics
```{r}
plot(fit.fb)
plot(fit.fb,which=4)
```
Everything looks good!






### GLM

Model Type: GLM, `family=Gamma(link='log')`

#### Model Selection: BIC

Backward Selection

```{r}
n <- dim(abund_dat)[1]
glmfit.a <- glm(mapy_mean_density ~ site + urchin_presence + urchin_abundance +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp +
              prev_mean_density + site:prev_mean_density, data=abund_dat, family=Gamma(link='log'))

glmfit.f <- step(glmfit.a, direction = "backward", k=log(n))
```
<br>

Forward Selection

```{r}
n <- dim(abund_dat)[1]
glmfit.e <- glm(mapy_mean_density ~ 1, data=abund_dat, family=Gamma(link='log'))
glmfit.ff <- step(glmfit.e, scope=scp, direction = "forward",  k = log(n))
```
<br>

Forward and backward selection chose the same model
```{r}
summary(glmfit.f)
plot(glmfit.f); plot(glmfit.f, which=4)
```
The residuals skew right.
```{r}
hist(glmfit.f$residuals)
```
<br>


#### Model Selection: AIC

Backward Selection

```{r}
glmfit.fb <- step(glmfit.a, direction = "backward", k=2)
```
<br>

Forward Selection

```{r}
glmfit.ff <- step(glmfit.e, scope=scp, direction = "forward",  k = 2)
```
<br>

Forward and backward selection chose different models. Compare their fits
```{r}
summary(glmfit.ff); summary(glmfit.fb)
```

What does the ANOVA say?
```{r}
anova(glmfit.ff,glmfit.fb, test="F")
```
<br>
<br>
<br>
<br>



## Kelp Size

*Response:* log(`mapy_mean_fronds`)

Interactions: site by previous mean density

Since there are some missing data in the kelp size categories, I'm going to select only the variables we are interested in at the moment and filter for missing data.

```{r}
size_dat <- dplyr::select(lmdat.t, mapy_mean_fronds,site, urchin_presence, urchin_abundance, prev_mean_fronds, mean_ammonia, mean_no2_no3, mean_po4, mean_tchl, mean_pon, summer_mean_temp, summer_var_temp) %>%
  filter(!is.na(mapy_mean_fronds))

size_dat$prev_mean_fronds[which(size_dat$prev_mean_fronds == "-Inf")] <- 0
size_dat$urchin_presence <- as.factor(size_dat$urchin_presence)
size_dat$site <- as.factor(size_dat$site)


colnames(size_dat)
```
<br>


### LM

Model Type: OLS, Log transformation of response

#### Model Selection: BIC

Backward model selection

```{r}
n <- dim(size_dat)[1]
fit.a <- lm(log(mapy_mean_fronds) ~ site + urchin_presence + urchin_abundance + 
              prev_mean_fronds + site:prev_mean_fronds +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp, data=size_dat)

fit.fb <- step(fit.a, direction = "backward",  k = log(n))
```
<br>

Forward Model Selection 

```{r}
fit.e <- lm(log(mapy_mean_fronds) ~ 1, data=size_dat)

scp <- list(lower = ~ 1,
           upper = ~ site + urchin_presence + urchin_abundance + 
              prev_mean_fronds + site:prev_mean_fronds +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp)

fit.ff <- step(fit.e, scope=scp, direction = "forward",  k = log(n))
```
<br>

Forward and backward selection produces the same model! Check out the fit
```{r}
summary(fit.fb)
```
<br>

Yeah so... it's a pretty terrible model.

<br>
<br>

#### Model Selection: AIC

Backward model selection

```{r}
fit.fb <- step(fit.a, direction = "backward",  k = 2)
```
<br>

Forward Model Selection 

```{r}
fit.ff <- step(fit.e, scope=scp, direction = "forward",  k = 2)
```
<br>

Forward and backward selection produces very different models. Since we already saw the model with just site as a predictor variable, Let's look at the larger model. 
```{r}
summary(fit.fb)
```
<br>

Still a pretty terrible model.


### GLM

Model Type: GLM, `family=Gamma(link='log')`

#### Model Selection: BIC

Backward Selection

```{r}
glmfit.a <- glm(mapy_mean_fronds ~ site + urchin_presence + urchin_abundance +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp +
              prev_mean_fronds + site:prev_mean_fronds, data=size_dat, family=Gamma(link='log'))

glmfit.f <- step(glmfit.a, direction = "backward", k=log(n))
```
<br>

Forward Selection

```{r}
glmfit.e <- glm(mapy_mean_fronds ~ 1, data=size_dat, family=Gamma(link='log'))
glmfit.ff <- step(glmfit.e, scope=scp, direction = "forward",  k = log(n))
```
<br>

Forward and backward selection chose the same model; very different from the OLS model. 
```{r}
summary(glmfit.f)
anova(glmfit.e,glmfit.f, test="F")
plot(glmfit.f); plot(glmfit.f, which=4)
```
There is non-constant variance of residuals according to the TA plot.

<br>


#### Model Selection: AIC

Backward Selection

```{r}
glmfit.fb <- step(glmfit.a, direction = "backward", k=2)
```
<br>

Forward Selection

```{r}
glmfit.ff <- step(glmfit.e, scope=scp, direction = "forward",  k = 2)
```
<br>

Forward and backward selection chose the same model, with one extra variable compared to the BIC - based models selection (`mean_no2_no3`). 
```{r}
summary(glmfit.ff)
anova(glmfit.e,glmfit.ff,test="F")
plot(glmfit.ff)
```

Maybe some non-constant variance, but overall it seems ok.
<br>
<br>
<br>
<br>
<br>



## Kelp Size (holdfasts)

*Response:* `mapy_mean_hld`

Interactions: site by previous mean size

Since there are some missing data in the kelp size categories, I'm going to select only the variables we are interested in at the moment and filter for no missing data.

```{r}
size_dat <- dplyr::select(lmdat.t, mapy_mean_hld,site, urchin_presence, urchin_abundance, prev_mean_hld, mean_ammonia, mean_no2_no3, mean_po4, mean_tchl, mean_pon, summer_mean_temp, summer_var_temp) %>%
  filter(!is.na(mapy_mean_hld))

size_dat$prev_mean_hld[which(is.na(size_dat$prev_mean_hld))] <- 0
size_dat$urchin_presence <- as.factor(size_dat$urchin_presence)
size_dat$site <- as.factor(size_dat$site)

colnames(size_dat)
```
<br>


### LM

Model Type: OLS

#### Model Selection: BIC

Backward model selection

```{r}
n <- dim(size_dat)[1]
fit.a <- lm(mapy_mean_hld ~ site + urchin_presence + urchin_abundance + 
              prev_mean_hld + site:prev_mean_hld +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp, data=size_dat)

fit.fb <- step(fit.a, direction = "backward",  k = log(n))
```
<br>

Forward Model Selection 

```{r}
fit.e <- lm(mapy_mean_hld ~ 1, data=size_dat)

scp <- list(lower = ~ 1,
           upper = ~ site + urchin_presence + urchin_abundance + 
              prev_mean_hld + site:prev_mean_hld +
              mean_ammonia + mean_no2_no3 + mean_po4 + mean_tchl + mean_pon +
              summer_mean_temp + summer_var_temp)

fit.ff <- step(fit.e, scope=scp, direction = "forward",  k = log(n))
```
<br>

Forward and backward selection produce the same models, only including the site variable. It's a pretty terrible model.
```{r}
summary(fit.ff)
```
<br>
```{r}
plot(fit.fb)
```

<br>
<br>

#### Model Selection: AIC

Backward model selection

```{r}
fit.fb <- step(fit.a, direction = "backward",  k = 2)
```
<br>

Forward Model Selection 

```{r}
fit.ff <- step(fit.e, scope=scp, direction = "forward",  k = 2)
```
<br>

Forward and backward selection produces  different models, one with more than just the site variable.
```{r}
summary(fit.fb); anova(fit.e,fit.ff,fit.fb,test="F")
```
<br>

The r-squared for the largest model is not much better than the models just looking at "site", but the normality of residuals is better!
```{r}
plot(fit.fb)
hist(fit.fb$residuals)
```
<br>
